for #2567, add RouteConditionEngine
