+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/core/strategy/route/value/BetweenRouteValue.java
@@ -20,6 +20,7 @@ package org.apache.shardingsphere.core.strategy.route.value;
 import com.google.common.collect.Range;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.ToString;
 
 /**
  * Route value for between.
@@ -28,6 +29,7 @@ import lombok.RequiredArgsConstructor;
  */
 @RequiredArgsConstructor
 @Getter
+@ToString
 public final class BetweenRouteValue<T extends Comparable<?>> implements RouteValue {
     
     private final String columnName;
